Validações parte 1 - Controle de fluxos

+++ app_locadora_carros/app/Http/Controllers/MarcaController.php
@@ -47,6 +47,10 @@ class MarcaController extends Controller
     {
         $marca = $this->marca->find($id);
 
+        if ($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe!'];
+        }
+
         return $marca;
     }
 
@@ -60,6 +64,11 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe!'];
+        }
+
         $marca->update($request->all());
 
         return $marca;
@@ -74,6 +83,11 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe!'];
+        }
+
         $marca->delete();
 
         return ['msg' => 'A marca foi removida com sucesso!'];
